fix: 26_BuilderPatternOrderApp のコードレビュー指摘を解消

+++ 26_BuilderPatternOrderApp/src/main/java/com/javalab/builderorder/OrderItem.java
@@ -10,9 +10,11 @@ public record OrderItem(String productName, int quantity, int unitPrice) {
 
     /**
      * 小計(数量×単価)を返す。
+     * int同士の乗算はオーバーフローしうる(例: 100000 * 100000は本来10000000000だが
+     * intでは1410065408になる)ため、long演算にしてから返す。
      * @return 小計
      */
-    public int subtotal() {
-        return quantity * unitPrice;
+    public long subtotal() {
+        return (long) quantity * unitPrice;
     }
 }
